requirements.txt
Flask==2.3.2
spotipy==2.22.1
python-dotenv==1.0.0
